Revert "TRY-33 yaml config einlesen (#8)"
This reverts commit c321de7b7bdc31eb7a67f802197242525a0d1842.

+++ src/scene/settings.rs
@@ -1,13 +1,5 @@
-use serde::{Deserialize, Serialize};
-
-// This struct represents settings for a program and is both serializable and deserializable
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct Settings {
-    
-    // This field specifies the maximum number of bounces for ray tracing or similar operations
-    max_bounces: usize,
-
-    // This field represents the number of samples used in rendering or some other operation
-    samples: usize,
+    max_bounces: u32,
+    samples: u32,
 }
-
